update try to fix cashier
