+++ app/(dashboard)/[storeId]/(routes)/cashier/components/cell-action.tsx
