+++ app/(dashboard)/[storeId]/(routes)/cashier/components/client.tsx
