+++ app/(dashboard)/[storeId]/(routes)/cashier/components/columns.tsx
